Remove the everest restart functionality
The restart functionality was added to support a research project.
However it turned out to be unnecessary for that project. Since
functionality should not be added for such a purpose only and the
general usefulness of the feature is unclear it should be removed.
