--- test-data/everest/math_func/config_restart.yml
@@ -1,95 +1,0 @@
-# This config file is the same as config_advanced.yml, except that it uses the
-# SciPy backend.
-
-wells: []
-
-controls:
-- initial_guess: 0.25
-  max: 1.0
-  min: -1.0
-  name: point
-  perturbation_magnitude: 0.005
-  type: generic_control
-  variables:
-  - name: x
-    index: 0
-  - name: x
-    index: 1
-  - name: x
-    index: 2
-
-objective_functions:
-  -
-    name: distance
-
-input_constraints:
-  -
-    weights:
-      point.x-0: 0
-      point.x-1: 0
-      point.x-2: 1
-    upper_bound: 0.4
-
-output_constraints:
-  - name: x-0_coord
-    lower_bound: 0.1
-    scale: 0.1
-
-
-# Optimal value expected at x=0.1, y=0, z=0.4, with distance 3.72
-
-
-install_jobs:
-  -
-    name: adv_distance3
-    source: jobs/ADV_DISTANCE3
-  -
-    name: adv_dump_controls
-    source: jobs/ADV_DUMP_CONTROLS
-
-
-forward_model:
- # Compute distance (squared and negated) between 2 points
-  - adv_distance3     --point-file point.json
-                      --target-file data/r{{ realization}}/target.json
-                      --out distance
- # Write the value of each control to a separate file
-  - adv_dump_controls --controls-file point.json
-                      --out-suffix _coord
-
-
-
-model:
-  realizations: [0, 2]
-  realizations_weights: [ 0.25, 0.75]
-
-
-install_data:
-  -
-    link: false
-    source: r{{ configpath }}/adv_target_r{{ realization }}.json
-    target: data/r{{ realization}}/target.json
-
-
-install_templates: []
-
-optimization:
-  backend: scipy
-  algorithm: SLSQP
-  convergence_tolerance: 0.001
-  constraint_tolerance: 0.001
-  perturbation_num: 4
-  speculative: True
-  max_function_evaluations: 3
-  restart:
-    max_restarts: 1
-    restart_from: "last"
-
-environment:
-  log_level: debug
-  random_seed: 123
-  simulation_folder: scratch/advanced/
-  output_folder: everest_output/
-
-simulator:
-  enable_cache: false
